requirements.txt
google-cloud-pubsub
google-cloud-bigquery
google-genai
pytchat
python-dotenv
fastapi
uvicorn
